+++ sdk/clojure/adapter-http-kit/src/main/starfederation/datastar/clojure/adapter/http_kit.clj
@@ -6,6 +6,8 @@
     [starfederation.datastar.clojure.utils :refer [def-clone]]))
 
 
+(def-clone on-open ac/on-open)
+(def-clone on-close ac/on-close)
 (def-clone on-exception ac/on-exception)
 (def-clone default-on-exception ac/default-on-exception)
 
@@ -18,26 +20,30 @@
 (def-clone gzip-buffered-writer-profile ac/gzip-buffered-writer-profile)
 
 
+;; TODO: remove deprecated get-on-open/close next release
 (defn ->sse-response
   "Make a Ring like response that will start a SSE stream.
 
   The status code and the the SSE specific headers are sent automatically
-  before `on-open` is called.
+  before [[on-open]] is called.
 
   Note that the SSE connection stays opened util you close it.
 
   General options:
-  - `:status`: Status for the HTTP response, defaults to 200
-  - `:headers`: Ring headers map to add to the response.
-  - `:on-open`: Mandatory callback `(fn [sse-gen] ...)` called when the
-    generator is ready to send.
-  - `:on-close`: callback `(fn [sse-gen status-code]...)` called when the
-    underlying Http-kit AsyncChannel is closed.
-  - [[on-exception]]: callback called when sending a SSE event throws
-  - [[write-profile]]: write profile for the connection
-    defaults to [[basic-profile]]
+  - `:on-open`: deprecated in favor of [[on-open]]
+  - `:on-close`: deprecated in favor of [[on-close]]
 
-  When it comes to write profiles, the SDK provides:
+  - `:status`: status for the HTTP response, defaults to 200.
+  - `:headers`: ring headers map to add to the response.
+  - [[on-open]]: mandatory callback called when the generator is ready to send.
+  - [[on-close]]: callback called when the underlying Http-kit AsyncChannel is
+    closed. It receives a second argument, the `:status-code` value we get from
+    the closing AsyncChannel.
+  - [[on-exception]]: callback called when sending a SSE event throws.
+  - [[write-profile]]: write profile for the connection.
+    Defaults to [[basic-profile]]
+
+  SDK provided write profiles:
   - [[basic-profile]]
   - [[buffered-writer-profile]]
   - [[gzip-profile]]
@@ -46,8 +52,11 @@
   You can also take a look at the `starfederation.datastar.clojure.adapter.common`
   namespace if you want to write your own profiles.
   "
-  [ring-request {:keys [on-open on-close] :as opts}]
-  (let [future-send! (promise)
+  [ring-request opts]
+  {:pre [(ac/get-on-open opts)]}
+  (let [on-open-cb (ac/get-on-open opts)
+        on-close-cb (ac/get-on-close opts)
+        future-send! (promise)
         future-gen (promise)]
     (hk-server/as-channel ring-request
       {:on-open
@@ -57,15 +66,15 @@
                sse-gen (impl/->sse-gen ch send! opts)]
            (deliver future-gen sse-gen)
            (deliver future-send! send!)
-           (on-open sse-gen)))
+           (on-open-cb sse-gen)))
 
        :on-close
        (fn [_ status]
          (let [closing-res
                (ac/close-sse!
                 #(when-let [send! (deref future-send! 0 nil)] (send!))
-                #(when on-close
-                   (on-close (deref future-gen 0 nil) status)))]
+                #(when on-close-cb
+                   (on-close-cb (deref future-gen 0 nil) status)))]
            (if (instance? Exception closing-res)
              (throw closing-res)
              closing-res)))})))