+++ sdk/clojure/adapter-ring/src/main/starfederation/datastar/clojure/adapter/ring/impl.clj
@@ -23,6 +23,7 @@
       (write! writer event-type data-lines event-opts)
       (ac/flush writer)))))
 
+;; TODO: Remove the get-on-open/close next release
 
 ;; Note that the send! field has 2 usages:
 ;; - it stores the sending function
@@ -43,15 +44,15 @@
 
     (let [!error (volatile! nil)]
       (try
-        ;; initializing the writing machinery
+        ;; initializing the internal state
         (let [opts (::opts response)]
           (set! send! (->send output-stream opts))
           (set! on-exception (or (ac/on-exception opts)
                                  ac/default-on-exception))
-          (when-let [cb (:on-close opts)]
+          (when-let [cb (ac/get-on-close opts)]
             (set! on-close cb)))
 
-        ;; flushing the HTTP headers
+        ;; flush the HTTP headers
         (.flush ^OutputStream output-stream)
         true ;; dummy return
         ;; We catch everything here, if not a Throwable may pass through
@@ -60,14 +61,14 @@
           (vreset! !error t))
         (finally
           ;; Any exception should have been caught,
-          ;; the setup for the writing machinery is done,
+          ;; the setup the internal state is done,
           ;; the HTTP headers are sent
           ;; we can now release the lock now
           (.unlock lock)
           (if-let [e @!error]
             (throw e) ;; if error throw, the lock is already released
             ;; if all is ok call on-open, it can safely throw...
-            (when-let [on-open (-> response ::opts :on-open)]
+            (when-let [on-open (-> response ::opts (ac/get-on-open))]
               (on-open this)))))))
  
   p/SSEGenerator